requirements.txt
pandas
scikit-learn
streamlit
Pillow
requests
